requirements.txt
Flask
pandas
networkx
matplotlib
gmplot
python-dotenv